Strip markdown from notification banner bodies
The banner showed the raw source, so a body written in markdown read as
syntax. MarkdownPreview already flattens it for previews; move it out of
Markdown.swift, which imports MarkdownUI and cannot compile in the
notification-service extension, and run both banner paths through it.
